chore(storybook): run e2es on macos
